type of service with enum and other changes
