+++ src/app/business.ts
@@ -41,3 +41,46 @@ export class Business {
     ){}
     }    
 
+
+export enum BusinessType{
+        STORAGESPACEPROVIDERS= "Storage Space Providers",
+        PANDITHSERVICES = "Pandith Services",
+        FASHIONDESIGNERSERVICES = "Fashion Designers Services",
+        FLEXIPRINTINGSERVICES = "Flex Printing Services",
+        GIFTSSHOP="Gifts Shop",
+        MATRIMONIALSERVICES="Matrimonial Services",
+        EVENTORGANIZERS="Event Organizers",
+        JEWELRYSTORES="Jewelery Stores",
+        BRIDALBEAUTYPRODUCTS = "Bridal Beauty Products",
+        ACOUSTICSINGERSBANDS="Acoustic Singer Bands",
+        BRIDALHAIRCARE="Bridal Haircare",
+        BRIDALSKINCARE="Bridal Skincare",
+        CAKESBAKES="Cakes n Bakes",
+        CATERING="Cartering",
+        GENERALCLEANINGSERVICES="General Cleaining Services",
+        CORPORATEEVENTVENUE="Corporate Event Venue",
+        DJSENTERTAINMENTAGENCY="DJ Entertainment Agency",
+        EVENTPHOTOGRAPHY="Event Photography",
+        EVENTSECURITY="Event Security",
+        FOODTRUCK="Food Truck",
+        GRAPHICDESIGNER="Graphic Designer",
+        FIREWORKS="Fireworks",
+        INVITATIONCARDPRINT="Invitation Card Printing",
+        KIDSPARTYHOST="Kids party Host",
+        MAGICIANS="Magicians",
+        MOBILEBAR="Mobile Bar",
+        PARTYEVENTCLEANING="Party Event Cleaning",
+        PETSITTING="Pet Sitting",
+        WEDDINGPHOTOGRAPHY="Wedding Photography",
+        WEDDINGTRANSPORTATION="Wedding Transportation",
+        WEDDINGVENUES="Wedding Venues",
+        WEDDINGSEVENTFLORISTS="Wedding Event Florists"
+}
+
+export class BusinessPasswordChange {
+    constructor(
+        public password: string,
+        public newPassword: string,
+        public confirmNewPassword: string
+    ){}
+}
