+++ src/app/find-business/find-business.component.ts
@@ -1,5 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { Router, ActivatedRoute } from '@angular/router';
+import { BusinessType } from '../business';
 
 @Component({
   selector: 'app-find-business',
@@ -15,4 +16,13 @@ export class FindBusinessComponent implements OnInit {
     this.setKeywordValue=param1;
   }
   
+  businesstypes=BusinessType;
+  b_type_keys() : Array<string> {
+    var keys = Object.keys(this.businesstypes);
+    return keys;
+  }
+  b_type_values(): Array<string> {
+    var values = Object.values(this.businesstypes);
+    return values;
+  }
 }
