+++ src/app/header/header.component.html
@@ -9,7 +9,7 @@
         <div id="navbarNavDropdown" class="navbar-collapse collapse">
             <ul class="navbar-nav mr-auto">
                 <li class="nav-item dropdown">
-                <div class="input-group ml-6 " *ngIf="router.url == ''">
+                <div class="input-group ml-6 " *ngIf="router.url == '/'">
                     <div class="input-group-append">
                         <button class="btn btn-outline-light" type="button" (click)="onSearch(searchValue.value)">
                             <em class="fa fa-search"></em>
